Cleanup code. New fetch process. All from CLGeocoder. Added Enum helper
