+++ Fowl Weather/Fowl Weather/Model Controllers/WeatherController.swift
@@ -19,83 +19,6 @@ class WeatherController {
     var fiveDayForcast: [ForcastedWeatherDay]?
     var weatherDays: [ForcastedWeatherDay] = []
     
-    //MARK: - Fetch Current Weather By City Name
-    
-    func fetchWeatherByCityName(cityName: String, completion: @escaping (CurrentWeather?, Error?) -> Void ) {
-        var url = baseURL.appendingPathComponent("weather")
-        
-        var components = URLComponents(url: url, resolvingAgainstBaseURL: true)!
-        let cityQueryItem = URLQueryItem(name: "q", value: cityName)
-        let imperialQueryItem = URLQueryItem(name: "units", value: "imperial")
-        let apiKeyQueryItem = URLQueryItem(name: "apiKey", value: apiKey)
-        components.queryItems = [cityQueryItem, imperialQueryItem, apiKeyQueryItem]
-        url = components.url!
-        print("The url is \(url)")
-        
-        Group.dispatchGroup.enter()
-        URLSession.shared.dataTask(with: url) { (data, _, error) in
-            
-            if let error = error {
-                NSLog("Error fetching city by name:\(error)")
-                completion(nil, error)
-                return
-            }
-            
-            guard let data = data else {
-                let error = NSError()
-                NSLog("Bad data. Error with date when fetching weather by cityname: \(cityName) with error:\(error)")
-                return
-            }
-            
-            do{
-                let weatherDay = try JSONDecoder().decode(CurrentWeather.self, from: data)
-                self.currentWeather = weatherDay
-                completion(self.currentWeather, nil)
-            } catch {
-                NSLog("Unable to decode data into CurrentWeather object:\(error)")
-            }
-            Group.dispatchGroup.leave()
-        }.resume()
-    }
-    
-    //MARK: - Fetch Current Weather By Zip Code
-    
-    func fetchWeatherByZipCode(zipCode: String, completion: @escaping (CurrentWeather?, Error?) -> Void ) {
-        
-        var url = baseURL.appendingPathComponent("weather")
-        var components = URLComponents(url: url, resolvingAgainstBaseURL: true)!
-        
-        let zipQueryItem = URLQueryItem(name: "zip", value: zipCode)
-        let imperialQueryItem = URLQueryItem(name: "units", value: "imperial")
-        let apiKeyQueryItem = URLQueryItem(name: "apiKey", value: apiKey)
-        components.queryItems = [zipQueryItem, imperialQueryItem, apiKeyQueryItem]
-        url = components.url!
-        
-        Group.dispatchGroup.enter()
-        URLSession.shared.dataTask(with: url) { (data, _, error) in
-            if let error = error {
-                NSLog("Error fetching city by name:\(error)")
-                completion(nil, error)
-                return
-            }
-            
-            guard let data = data else {
-                let error = NSError()
-                NSLog("Bad data. Error with date when fetching weather by zipcode: \(zipCode) with error:\(error)")
-                return
-            }
-            
-            do{
-                let weatherDay = try JSONDecoder().decode(CurrentWeather.self, from: data)
-                self.currentWeather = weatherDay
-                completion(self.currentWeather, nil)
-            } catch {
-                NSLog("Unable to decode data into CurrentWeather object:\(error)")
-            }
-            Group.dispatchGroup.leave()
-        }.resume()
-    }
-    
     //MARK: - Fetch Current By User Location
     
     func fetchWeatherByLocation(location: CLLocation, completion: @escaping (CurrentWeather?, Error?) -> Void ) {
@@ -135,86 +58,7 @@ class WeatherController {
              Group.dispatchGroup.leave()
          }.resume()
      }
-    
-    //MARK: - Fetch 5-Day Weather Forcast By City Name
-    
-    func fetchFiveDayByCityName(cityName: String, completion: @escaping ([ForcastedWeatherDay]?, Error?) -> Void ) {
-        var url = baseURL.appendingPathComponent("forecast")
-        
-        var components = URLComponents(url: url, resolvingAgainstBaseURL: true)!
-        let cityQueryItem = URLQueryItem(name: "q", value: cityName)
-        let imperialQueryItem = URLQueryItem(name: "units", value: "imperial")
-        let apiKeyQueryItem = URLQueryItem(name: "apiKey", value: apiKey)
-        components.queryItems = [cityQueryItem, imperialQueryItem, apiKeyQueryItem]
-        url = components.url!
-        print("The url is \(url)")
-        
-        Group.dispatchGroup.enter()
-        URLSession.shared.dataTask(with: url) { (data, _, error) in
-            if let error = error {
-                NSLog("Error fetching city by name:\(error)")
-                completion(nil, error)
-                return
-            }
-            
-            guard let data = data else {
-                let error = NSError()
-                NSLog("Bad data. Error with date when fetching weather by cityname: \(cityName) with error:\(error)")
-                return
-            }
-            
-            do{
-                let weatherList = try JSONDecoder().decode(FiveDayForcast.self, from: data)
-                self.weatherDays = weatherList.list
-                self.fiveDayForcast = self.cleanFiveDayForecast(daysToClean: self.weatherDays)
-                completion(self.weatherDays, nil)
-            } catch {
-                NSLog("Unable to decode data into CurrentWeather object:\(error)")
-            }
-            Group.dispatchGroup.leave()
-        }.resume()
-    }
-    
-    //MARK: - Fetch 5-Day Weather Forcast By Zip
-    
-    func fetchFiveDayByZipCode(zipCode: String, completion: @escaping ([ForcastedWeatherDay]?, Error?) -> Void ) {
-        var url = baseURL.appendingPathComponent("forecast")
-        
-        var components = URLComponents(url: url, resolvingAgainstBaseURL: true)!
-        let cityQueryItem = URLQueryItem(name: "zip", value: zipCode)
-        let imperialQueryItem = URLQueryItem(name: "units", value: "imperial")
-        let apiKeyQueryItem = URLQueryItem(name: "apiKey", value: apiKey)
-        components.queryItems = [cityQueryItem, imperialQueryItem, apiKeyQueryItem]
-        url = components.url!
-        print("The url is \(url)")
-        
-        Group.dispatchGroup.enter()
-        URLSession.shared.dataTask(with: url) { (data, _, error) in
-            
-            if let error = error {
-                NSLog("Error fetching city by name:\(error)")
-                completion(nil, error)
-                return
-            }
-            
-            guard let data = data else {
-                let error = NSError()
-                NSLog("Bad data. Error with date when fetching weather by zip code: \(zipCode) with error:\(error)")
-                return
-            }
 
-            do{
-                let weatherList = try JSONDecoder().decode(FiveDayForcast.self, from: data)
-                self.weatherDays = weatherList.list
-                self.fiveDayForcast = self.cleanFiveDayForecast(daysToClean: self.weatherDays)
-                completion(self.fiveDayForcast, nil)
-            } catch {
-                NSLog("Unable to decode data into CurrentWeather object:\(error)")
-            }
-            Group.dispatchGroup.leave()
-        }.resume()
-    }
-    
     //MARK: - Fetch 5-Day Weather Forcast By Location
     
     func fetchFiveDayByLocation(location: CLLocation, completion: @escaping ([ForcastedWeatherDay]?, Error?) -> Void ) {
