--- Fowl Weather/Fowl Weather/Views/AlexsCustomView.swift
@@ -1,9 +1,0 @@
-//
-//  AlexsCustomView.swift
-//  Fowl Weather
-//
-//  Created by Jake Connerly on 11/18/19.
-//  Copyright © 2019 jake connerly. All rights reserved.
-//
-
-import Foundation
